requirements.txt
# Agent Lightning Requirements

# Core dependencies
numpy>=1.21.0
torch>=2.0.0
gymnasium>=0.28.0

# API and Web
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
websockets>=11.0
grpcio>=1.50.0

# Monitoring and Dashboard
streamlit>=1.25.0
plotly>=5.14.0
pandas>=2.0.0

# Machine Learning
scikit-learn>=1.3.0
ray[rllib]>=2.5.0

# Authentication and Security
pyjwt>=2.8.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# Database and Caching
redis>=4.5.0
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.23
alembic>=1.13.0
asyncpg>=0.29.0
python-dotenv>=1.0.0

# Observability
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
opentelemetry-exporter-otlp>=1.20.0
opentelemetry-instrumentation-requests>=0.41b0
opentelemetry-instrumentation-logging>=0.41b0
prometheus-client>=0.17.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0

# Optional (for cloud deployment)
boto3>=1.28.0  # AWS
azure-storage-blob>=12.17.0  # Azure
google-cloud-storage>=2.10.0  # GCP